requirements.deploy.txt
azureml-pipeline
python-dotenv